Move ServeMux into seperate file
This reduces the clutter in server.go.
